requirements.txt
msal
requests
psycopg2-binary
python-dotenv
uuid
schedule
boto3
fastapi
uvicorn[standard]
